structure, comments and postman tests (local on my machine)
